example(agentic): simplify api
