+++ examples/nostr-agentic-example-application/readme.md
@@ -18,7 +18,7 @@ API docs at http://localhost:8080/swagger-ui/.
 
 ### `listmodels`
 ```shell
-curl --silent http://localhost:8080/api/v1/agentic/listmodels | jq
+curl --silent http://localhost:8080/api/v1/model/listmodels | jq
 ```
 ```json
 {
@@ -47,7 +47,7 @@ curl --silent http://localhost:8080/api/v1/agentic/listmodels | jq
 ### `nostr/event/plain`
 ```shell
 curl --silent --request 'POST' \
-    http://localhost:8080/api/v1/agentic/nostr/event/plain \
+    http://localhost:8080/api/v1/nostr/event-plain \
     --header 'Content-Type: application/json' \
     --data '{
       "contents": "What day is today?"
